[BUILD] Created libantares-core-checks
